pyproject.toml
[tool.poetry]
name = "fastapi-tortoise"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi="*"
tortoise-orm = "*"
aerich ="*"
python-multipart = "^0.0.5"
rearq = "^0.1.3"
fastapi-admin = "^0.3.2"

[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "^19.10b0"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"
pylint = "^2.6.0"

[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
